% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{pca}
\usage{
pca(x, threshold = 0.8, plot = FALSE)
}
\arguments{
\item{x}{(dataframe) The dataframe of input features}

\item{threshold}{(numeric) The correlation value that you wish to use as the
       cutoff. Feature pairs with a correlation value above this threshold
       will be returned.

       (DEFAULT = 0.8)}

\item{plot}{(logical) Should it print the highly correlated pairs?

       (DEFAULT = FALSE)}
}
\value{
A matrix, where each row contains a pair of column indices from x
        that are highly correlated with each other.
}
\description{
Principle Component Analysis. Compares all the features (columns) for some
dataset, and calculates which pairs of features are highly correlated with
each other. You can specify what threshold to use as your definition of
"highly correlated". You can also plot all the highly correlated pairs.
}
\examples{
library(kernlab);
data(spam)
pca(spam[,-68], threshold=0.7, plot=TRUE)
}
\author{
Ronny Restrepo
}

